Issue #2 done: User do not see slots they already chose on Make Reservation
On the "Make Reservation" or "Reservation Update" page, the list of available slots do not include the slots for which the invited user has already made a reservation. This will prevent the user from attempting to make duplicate reservations for a single slot.
In contrast to what I initially thought, the invited user is not able to reserve the same slot more than once; the web tool merely displays the "SUCCESS!!" page when the same slot is clicked again. This fix will prevent this issue altogether by not showing the already reserved slots at all.